requirements.txt
beautifulsoup4==4.11.1
Flask==2.0.3
matplotlib==3.5.2
mplfinance==0.12.9b0
numpy==1.22.4
opencv==4.5.5
pandas==1.4.2
pandas-datareader==0.10.0
Pillow==9.1.1
plotly==5.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.1.1
tqdm==4.64.0
typer==0.4.1
yfinance==0.1.70
